requirements.txt
Django==2.0.2
django-cors-headers==2.2.0
django-rest-swagger==2.1.2
djangorestframework==3.7.7
mysqlclient==1.3.13
requests==2.19.1
urllib3==1.23
djangorestframework-jwt==1.11.0
requests==2.19.1
simplejson==3.16.0
urllib3==1.23
psycopg2-binary==2.7.4
